Split CCL into triton/ and gluon/ subfolders
Move kernel definitions out of the top-level CCL files into
iris/ccl/triton/ and iris/ccl/gluon/ subpackages, mirroring
the existing iris/mem/ split. Each top-level file becomes a
thin dispatcher that delegates to the appropriate backend
based on config.use_gluon.

This eliminates the `if GLUON_AVAILABLE:` guards around kernel
definitions and the repeated try/except import blocks. Gluon
availability is now determined once in iris/ccl/gluon/__init__.py.

User-facing imports (from iris.ccl import Config, ReduceOp) and
the dispatch API (shmem.ccl.all_gather, etc.) are unchanged.
